cmd/agent перенести функции для парсинга структур в structs.go
